requirements.txt
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.0
botocore==1.35.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
fonttools==4.53.1
htmlmin==0.1.12
idna==3.7
ImageHash==4.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
mariadb==1.1.10
MarkupSafe==2.1.5
matplotlib==3.9.2
multimethod==1.12
networkx==3.3
numba==0.60.0
numpy==1.26.4
pandas==2.2.2
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
py4j==0.10.9.7
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pyspark==3.5.2
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.7.0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
scipy==1.13.1
seaborn==0.13.2
setuptools==72.1.0
soupsieve==2.6
statsmodels==0.14.2
tqdm==4.66.5
typeguard==4.3.0
tzdata==2024.1
urllib3==2.2.2
visions==0.7.6
wheel==0.43.0
wordcloud==1.9.3
ydata-profiling==4.9.0
